backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai==1.51.0
Pillow==10.4.0
python-multipart==0.0.12
httpx==0.27.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
